energyproj/requirements.txt
altgraph
appdirs
bottle
bottle-websocket
certifi
chardet
charset-normalizer
click
DateTime
decorator
distlib
Eel
filelock
future
geocoder
gevent
gevent-websocket
greenlet
idna
importlib-metadata
itsdangerous
macholib
MarkupSafe
modulegraph
numpy
psutil
py-cpuinfo
pymongo[srv]
python-dateutil
python-dotenv
pytz
ratelim
requests
six
urllib3
uuid
virtualenv
websocket-client
Werkzeug
whichcraft
zipp
zope.event
zope.interface
dnspython
